Split views by app
